beauty_coin/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.4.1
    # via django
cffi==1.15.0
    # via cryptography
cryptography==35.0.0
    # via pyopenssl
django==3.2.9
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-filter
    #   djangorestframework
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
markdown==3.3.4
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyopenssl==21.0.0
    # via -r requirements.in
pytz==2021.3
    # via django
six==1.16.0
    # via pyopenssl
sqlparse==0.4.2
    # via django
werkzeug==2.0.2
    # via -r requirements.in
